backend/requirements.txt
fastapi==0.95.2
langchain==0.0.304
langchain-experimental==0.0.22
pgvector==0.1.8
psycopg2==2.9.6
python-dotenv==1.0.0
SQLAlchemy==2.0.13
starlette==0.27.0
tiktoken==0.4.0
openai
uvicorn
websockets